Rename movie search list
