+++ src/components/index.ts
@@ -5,5 +5,5 @@ export { default as MovieRating } from "./MovieRating";
 export { default as Pagination } from "./Pagination";
 export { default as Spinner } from "./Spinner/Spinner";
 export { default as MovieSearchError } from "./MovieSearchError";
-export { default as MoviesList } from "./MoviesList";
+export { default as MovieSearchList } from "./MovieSearchList";
 export { default as NoMoviesFound } from "./NoMoviesFound";
